flask/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.1.1
flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.28
typing-extensions==4.10.0
werkzeug==3.0.1
wtforms==3.1.2
zipp==3.17.0
